Separate poptrie4.c from poptrie.c for IPv4 routing table operations.
