+++ Makefile.am
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2015 Hirochika Asai <asai@jar.jp>
+# Copyright (c) 2015-2017 Hirochika Asai <asai@jar.jp>
 # All rights reserved.
 #
 
@@ -9,7 +9,7 @@ noinst_HEADERS = buddy.h
 
 bin_PROGRAMS = poptrie_test_basic poptrie_test_basic6
 lib_LTLIBRARIES = libpoptrie.la
-libpoptrie_la_SOURCES = poptrie.c poptrie6.c poptrie.h buddy.c buddy.h poptrie_private.h
+libpoptrie_la_SOURCES = poptrie.c poptrie4.c poptrie6.c poptrie.h buddy.c buddy.h poptrie_private.h
 
 poptrie_test_basic_SOURCES = tests/basic.c
 poptrie_test_basic_LDADD = libpoptrie.la
