+++ configure.ac
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2015 Hirochika Asai <asai@jar.jp>
+# Copyright (c) 2015,2017 Hirochika Asai <asai@jar.jp>
 # All rights reserved.
 #
